+++ catalog/src/main/java/it/eng/catalog/model/Constraint.java
@@ -1,5 +1,7 @@
 package it.eng.catalog.model;
 
+import java.io.Serializable;
+
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
@@ -12,11 +14,13 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
-@EqualsAndHashCode
+@EqualsAndHashCode   // requires for offer check in negotiation flow
 @JsonDeserialize(builder = Constraint.Builder.class)
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class Constraint {
+public class Constraint implements Serializable {
 	
+	private static final long serialVersionUID = 1L;
+
 	@JsonProperty(DSpaceConstants.ODRL_LEFT_OPERAND)
 	private LeftOperand leftOperand;