+++ catalog/src/main/java/it/eng/catalog/rest/api/OfferAPIController.java
@@ -41,7 +41,6 @@ public class OfferAPIController {
         
         return ResponseEntity.badRequest().contentType(MediaType.APPLICATION_JSON)
         		.body(GenericApiResponse.error("Offer not valid"));
-
     }
 
 }