+++ negotiation/src/main/java/it/eng/negotiation/model/ContractNegotiation.java
@@ -1,8 +1,20 @@
 package it.eng.negotiation.model;
 
-import com.fasterxml.jackson.annotation.*;
+import java.util.Set;
+import java.util.stream.Collectors;
+
+import org.springframework.data.annotation.Id;
+import org.springframework.data.mongodb.core.mapping.DBRef;
+import org.springframework.data.mongodb.core.mapping.Document;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
+
 import it.eng.tools.model.DSpaceConstants;
 import jakarta.validation.ConstraintViolation;
 import jakarta.validation.Validation;
@@ -11,13 +23,6 @@ import jakarta.validation.constraints.NotNull;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.apache.commons.lang3.builder.ToStringBuilder;
-import org.springframework.data.annotation.Id;
-import org.springframework.data.mongodb.core.mapping.DBRef;
-import org.springframework.data.mongodb.core.mapping.Document;
-
-import java.util.Set;
-import java.util.stream.Collectors;
 
 /*
 {
@@ -104,13 +109,13 @@ public class ContractNegotiation extends AbstractNegotiationObject {
         	return this;
         }
 
-        @JsonSetter(DSpaceConstants.DSPACE_PROVIDER_PID)
+        @JsonProperty(DSpaceConstants.DSPACE_PROVIDER_PID)
         public Builder providerPid(String providerPid) {
             message.providerPid = providerPid;
             return this;
         }
 
-        @JsonSetter(DSpaceConstants.DSPACE_CONSUMER_PID)
+        @JsonProperty(DSpaceConstants.DSPACE_CONSUMER_PID)
         public Builder consumerPid(String consumerPid) {
             message.consumerPid = consumerPid;
             return this;
@@ -118,7 +123,7 @@ public class ContractNegotiation extends AbstractNegotiationObject {
 
         public ContractNegotiation build() {
             if (message.id == null) {
-                message.id = message.createNewId();
+                message.id = message.createNewPid();
             }
             if (message.providerPid == null) {
                 message.providerPid = message.createNewPid();
@@ -134,11 +139,6 @@ public class ContractNegotiation extends AbstractNegotiationObject {
                             .map(v -> v.getPropertyPath() + " " + v.getMessage())
                             .collect(Collectors.joining(", ")));
         }
-    }
-
-    @Override
-    public String toString() {
-        return ToStringBuilder.getDefaultStyle().toString();
     }
 
     @Override