+++ negotiation/src/test/java/it/eng/negotiation/model/ConstraintTest.java
@@ -1,9 +1,13 @@
 package it.eng.negotiation.model;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import it.eng.negotiation.serializer.Serializer;
 
 public class ConstraintTest {
 
@@ -26,6 +30,24 @@ public class ConstraintTest {
 				.build();
 		assertFalse(constraintA.equals(constraintB));
 		
+	}
+	
+	@Test
+	@DisplayName("Plain serialize/deserialize")
+	public void equalsTestPlain() {
+		Constraint constraint = MockObjectUtil.CONSTRAINT;
+		String ss = Serializer.serializePlain(constraint);
+		Constraint obj = Serializer.deserializePlain(ss, Constraint.class);
+		assertThat(constraint).usingRecursiveComparison().isEqualTo(obj);
+	}
+	
+	@Test
+	@DisplayName("Protocol serialize/deserialize")
+	public void equalsTestProtocol() {
+		Constraint constraint = MockObjectUtil.CONSTRAINT;
+		String ss = Serializer.serializeProtocol(constraint);
+		Constraint obj = Serializer.deserializeProtocol(ss, Constraint.class);
+		assertThat(constraint).usingRecursiveComparison().isEqualTo(obj);
 	}
 }