+++ negotiation/src/test/java/it/eng/negotiation/model/ContractAgreementMessageTest.java
@@ -1,5 +1,6 @@
 package it.eng.negotiation.model;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -96,6 +97,22 @@ public class ContractAgreementMessageTest {
 	public void missingContextAndType() {
 		JsonNode result = Serializer.serializePlainJsonNode(contractAgreementMessage);
 		assertThrows(ValidationException.class, () -> Serializer.deserializeProtocol(result, ContractAgreementMessage.class));
+	}
+	
+	@Test
+	@DisplayName("Plain serialize/deserialize")
+	public void equalsTestPlain() {
+		String ss = Serializer.serializePlain(contractAgreementMessage);
+		ContractAgreementMessage obj = Serializer.deserializePlain(ss, ContractAgreementMessage.class);
+		assertThat(contractAgreementMessage).usingRecursiveComparison().isEqualTo(obj);
+	}
+	
+	@Test
+	@DisplayName("Protocol serialize/deserialize")
+	public void equalsTestProtocol() {
+		String ss = Serializer.serializeProtocol(contractAgreementMessage);
+		ContractAgreementMessage obj = Serializer.deserializeProtocol(ss, ContractAgreementMessage.class);
+		assertThat(contractAgreementMessage).usingRecursiveComparison().isEqualTo(obj);
 	}
 	
 	private void validateAgreementProtocol(JsonNode agreement) {